+++ Changelog.md
@@ -11,6 +11,9 @@
 - New `coq.option.available?`
 - New `coq.bind-ind-parameters`
 
+### APPS
+- New `locker` app providing `lock` and `mlock` commands
+
 ## [1.11.2] - 24-09-2021
 
 Requires Elpi 1.13.6 and Coq 8.14.